+++ internal/vault/secrets/gcpkms/gcpkms_decrypt_ephemeral_resource.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package ephemeralsecrets
+package gcpkms
 
 import (
 	"context"
